SkillRepository interface created; SkillController mapped

+++ src/main/java/co/uk/bransby/equinetrainingtrackerapi/persistence/SkillRepository.java
@@ -2,6 +2,7 @@ package co.uk.bransby.equinetrainingtrackerapi.persistence;
 
 import co.uk.bransby.equinetrainingtrackerapi.models.Skill;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
-public interface SkillRepository extends JpaRepository<Skill, Integer> {
+public interface SkillRepository extends JpaRepository<Skill, Long> {
 }
